The page width is a layout quantity, and the sample menu has a test (ovation#318)
The push gate refused Double in the app's sources: floating point is forbidden on
the money path and CGFloat is deliberately exempt as a layout quantity
(scripts/check-forbidden-constructs.sh). A page width is layout.

The test gate then refused the Debug sample menu as untested, correctly. Its
cases: an ordinary sample shows, a refused one opens nothing and says the gate's
own sentence rather than a second wording of it, pressing on clears the refusal
so it cannot sit over the next sample (L243), closing puts both away, and every
sample in the menu can actually be built and says something of its own, so the
next sample added to the menu and not to the world fails here rather than on
screen (L546).

+++ Ovation/Document/ReviewSheetPresenter.swift
@@ -1,3 +1,4 @@
+import CoreGraphics
 import Foundation
 
 /// What the review sheet says, decided here rather than in a view (ovation#318).
@@ -56,10 +57,14 @@ final class ReviewSheetPresenter {
     /// rather than written down beside it: a review screen may not re-typeset the
     /// page to fit a column, so the only thing that changes is how large it is
     /// (PRD 52b, 10c).
-    var pageWidth: Double { (Self.pagePoints * scale.factor).rounded() }
+    /// CGFloat, NOT Double, and that is the app's rule rather than a preference:
+    /// floating point is forbidden on the money path and CGFloat is deliberately
+    /// exempt as a LAYOUT quantity (scripts/check-forbidden-constructs.sh). A page
+    /// width is layout.
+    var pageWidth: CGFloat { (Self.pagePoints * CGFloat(scale.factor)).rounded() }
 
     /// A US Letter page, which is what InvoicePDF draws.
-    private static let pagePoints: Double = 612
+    private static let pagePoints: CGFloat = 612
 
     /// How many times the session has rendered, for the tests that hold this to one
     /// render however many times the page is shown (PRD 10c).
